build: split manifest and code example builds
