+++ libs/components/manifest-generator/src/index.ts
@@ -1,1 +1,3 @@
 export { generateManifest } from './lib/generate-manifest.js';
+export { generateCodeExamplesManifest } from './lib/generate-code-examples-manifest.js';
+export { getComponentProjectNames } from './lib/utility/get-component-project-names.js';
