+++ libs/components/manifest-generator/src/lib/generate-manifest.ts
@@ -1,43 +1,23 @@
 import type {
-  SkyManifestCodeExamples,
   SkyManifestDocumentationConfig,
   SkyManifestPublicApi,
 } from '@skyux/manifest-local';
 
-import fs from 'node:fs';
 import fsPromises from 'node:fs/promises';
 import path from 'node:path';
 
-import { getCodeExamples } from './get-code-examples.js';
 import { getDocumentationConfig } from './get-documentation-config.js';
 import { getProjectDefinitions } from './get-project-definitions.js';
 import { getPublicApi } from './get-public-api.js';
-
-interface SkyManifestOptions {
-  codeExamplesPackageName: string;
-  outDir: string;
-  projectNames: string[];
-  projectsRootDirectory: string;
-}
-
-async function ensureDirectory(directoryPath: string): Promise<void> {
-  if (!fs.existsSync(directoryPath)) {
-    await fsPromises.mkdir(directoryPath);
-  }
-}
+import { SkyManifestOptions } from './manifest-options.js';
+import { ensureDirectory, getOutputPaths } from './utility/get-output-paths.js';
 
 async function writeManifestFiles(
   outDir: string,
   publicApi: SkyManifestPublicApi,
   documentationConfig: SkyManifestDocumentationConfig,
-  codeExamples: SkyManifestCodeExamples,
 ): Promise<void> {
-  const publicApiPath = path.join(outDir, 'public-api.json');
-  const documentationConfigPath = path.join(
-    outDir,
-    'documentation-config.json',
-  );
-  const codeExamplesPath = path.join(outDir, 'code-examples.json');
+  const { publicApiPath, documentationConfigPath } = getOutputPaths(outDir);
 
   await ensureDirectory(outDir);
 
@@ -54,18 +34,11 @@ async function writeManifestFiles(
   );
 
   console.log(`Created ${documentationConfigPath}`);
-
-  await fsPromises.writeFile(
-    codeExamplesPath,
-    JSON.stringify(codeExamples, undefined, 2),
-  );
-
-  console.log(`Created ${codeExamplesPath}\n`);
 }
 
 /**
  * Generates manifest files for the distribution build.
- * (This function is executed by the postbuild script.)
+ * (This function is executed by the manifest build script.)
  */
 export async function generateManifest(
   options: SkyManifestOptions,
@@ -80,17 +53,7 @@ export async function generateManifest(
   const [documentationConfig, documentationConfigErrors] =
     await getDocumentationConfig(publicApi, projects);
 
-  const [codeExamples, codeExamplesErrors] = await getCodeExamples(
-    publicApi,
-    documentationConfig,
-    options.codeExamplesPackageName,
-  );
-
-  const errors = [
-    ...publicApiErrors,
-    ...documentationConfigErrors,
-    ...codeExamplesErrors,
-  ];
+  const errors = [...publicApiErrors, ...documentationConfigErrors];
 
   if (errors.length > 0) {
     throw new Error(
@@ -101,12 +64,7 @@ export async function generateManifest(
 
   const outDir = path.normalize(options.outDir);
 
-  await writeManifestFiles(
-    outDir,
-    publicApi,
-    documentationConfig,
-    codeExamples,
-  );
+  await writeManifestFiles(outDir, publicApi, documentationConfig);
 
   return { publicApi };
 }
