+++ libs/components/manifest-generator/src/lib/manifest-options.ts
@@ -1,0 +1,6 @@
+export interface SkyManifestOptions {
+  codeExamplesPackageName: string;
+  outDir: string;
+  projectNames: string[];
+  projectsRootDirectory: string;
+}
