+++ libs/components/manifest-generator/src/lib/utility/get-component-project-names.ts
@@ -1,0 +1,10 @@
+import { getProjects, workspaceRoot } from '@nx/devkit';
+
+import { FsTree } from 'nx/src/generators/tree.js';
+
+export function getComponentProjectNames(): string[] {
+  const tree = new FsTree(workspaceRoot, false);
+  return Array.from(getProjects(tree).entries())
+    .filter(([, project]) => project.tags?.includes('component'))
+    .map(([name]) => name);
+}
