+++ libs/components/manifest-generator/src/lib/utility/get-output-paths.ts
@@ -1,0 +1,28 @@
+import fs from 'node:fs';
+import fsPromises from 'node:fs/promises';
+import path from 'node:path';
+
+export function getOutputPaths(outDir: string): {
+  publicApiPath: string;
+  documentationConfigPath: string;
+  codeExamplesPath: string;
+} {
+  const publicApiPath = path.join(outDir, 'public-api.json');
+  const documentationConfigPath = path.join(
+    outDir,
+    'documentation-config.json',
+  );
+  const codeExamplesPath = path.join(outDir, 'code-examples.json');
+
+  return {
+    publicApiPath,
+    documentationConfigPath,
+    codeExamplesPath,
+  };
+}
+
+export async function ensureDirectory(directoryPath: string): Promise<void> {
+  if (!fs.existsSync(directoryPath)) {
+    await fsPromises.mkdir(directoryPath);
+  }
+}
