+++ libs/components/manifest/project.json
@@ -22,6 +22,16 @@
       }
     },
     "build": {
+      "inputs": ["^buildInputs", "buildInputs", "documentationInputs"],
+      "executor": "nx:run-commands",
+      "outputs": ["{options.outputPath}"],
+      "options": {
+        "command": "npx vite-node ./scripts/build-manifest.mts",
+        "outputPath": "dist/libs/components/manifest"
+      },
+      "dependsOn": ["^build", "prebuild"]
+    },
+    "postbuild": {
       "inputs": [
         "^buildInputs",
         "buildInputs",
@@ -34,7 +44,7 @@
         "command": "npx vite-node ./scripts/postbuild-manifest.mts",
         "outputPath": "dist/libs/components/manifest"
       },
-      "dependsOn": ["^build", "prebuild"]
+      "dependsOn": ["^build", "build"]
     },
     "lint": {
       "executor": "@nx/eslint:lint",
