+++ nx.json
@@ -147,6 +147,7 @@
   "namedInputs": {
     "buildInputs": [
       "{projectRoot}/**/*",
+      "{workspaceRoot}/scripts/build-{projectName}.ts",
       "{workspaceRoot}/scripts/postbuild-{projectName}.ts",
       "{workspaceRoot}/scripts/utils/**",
       "!{projectRoot}/**/*.spec.[jt]s",
